Refactoring, fixes and updates: Refactor major codes and files as well as updates
